created delete van endpoint

+++ back-end/controllers/van.controller.js
@@ -152,15 +152,18 @@ const patchVan = async (req, res)=>{
 
 const deleteVanById = async (req, res)=>{
   const {id}=req.params
-  
+ 
   if (!mongoose.isValidObjectId(id)) {
     return res.status(400).json({ msg: "bad request" });
   }
+  const vanById = await Van.findById(id);
+
+  if (!vanById) {
+    return res.status(404).json({ error: "request not found" });
+  }
   const vanToDelete = await Van.findOneAndDelete(id);
-  if (!vanToDelete) {
-    return res.status(404).json({ msg: 'request not found' });
-}
-res.status(200).send({deletedVan:vanToDelete})
+
+res.status(200).send({deletedVan:vanToDelete, message:'van deleted'})
  
   }
